pyproject.toml
[project]
name = "deubot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=21.0",
    "openai>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.0",
    "black>=23.0",
    "flake8>=6.0",
]
